requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.7
django-autoslug==1.9.8
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-recaptcha3==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
marshmallow==3.19.0
mutagen==1.46.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
